requirements.txt
Django==1.8
django-rest-swagger==0.3.5
djangorestframework==3.3.2
MySQL-python==1.2.5
PyYAML==3.11
